% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_news.R
\name{clean_news}
\alias{clean_news}
\title{Clean retrieved news articles}
\usage{
clean_news(data, min_nchar = 300, as_date = TRUE, drop_vars = TRUE,
  to_lower = TRUE, distinct = TRUE, drop_na = FALSE,
  tif_corpus = FALSE)
}
\arguments{
\item{data}{Tbl, returned from \code{\link{get_news}}.}

\item{min_nchar}{Integer, specifying the minimum number of characters of
articles to be kept in the corpus.}

\item{as_date}{Logical, indicating whether dates should be transformed to
class "Date".}

\item{drop_vars}{Logical, indicating whether all variables (other than
\code{title, text, discoverDate,} & \code{website.domainName}) should be
dropped. The Newsriver API (typically) returns 26 variables, many of which
contain sparse metadata.}

\item{to_lower}{Logical, indicating whether the \code{title} and \code{text}
variables should be transformed to lowercase.}

\item{distinct}{Logical, indicating whether only articles with either
distinct \code{title} or \code{text} values should be kept.}

\item{drop_na}{Logical, indicating whether to drop rows containing missing
values.}

\item{tif_corpus}{Logical, indicating whether the tibble should be a
\href{https://github.com/ropensci/tif}{TIF} valid corpus.}
}
\description{
\code{clean_news} wrangles the messy data fetched by \code{\link{get_news}},
returning a tidy tibble with sensible defaults.
}
\examples{
\dontrun{
clean_news(data = my_tbl)

clean_news(my_tbl, min_nchar = 500, tif_corpus = TRUE)
}

}
