% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_creds.R
\name{store_creds}
\alias{store_creds}
\alias{store_creds_temp}
\title{Store user credentials}
\usage{
store_creds()

store_creds_temp()
}
\description{
\code{store_creds} prompts you for a Newsriver API token and user agent and
\emph{permanently} stores them so they are available for \emph{every} R
session.

\code{store_creds_temp} prompts you for a Newsriver API token and user agent
and \emph{temporarily} stores them so they are available for only the
\emph{current} R session.
}
\section{Token registration}{


To register for an API token, visit the
\href{https://console.newsriver.io/api-token}{Newsriver site} and follow the
sign up procedures. Registration is free.
}

\section{Setting a user agent}{


The user agent is a string used to identify the client. This allows Newsriver
to identify who is using the API (and is important if something goes wrong).
A good default user agent is your email address. \strong{Note:} the
user agent you supply does not need to be wrapped in quotes (nor does the
API token).
}

\section{Environment variables}{


Both \code{store_creds} and \code{store_creds_temp} store the supplied API
token and user agent as environment variables called "NEWSRIVER_API_KEY" and
"NEWSRIVER_USER_AGENT" respectively. These environment variables are then
called by other newsrivr functions.

To store these environment variables, \code{store_creds} tries to find a
.Renviron file located in your home directory. If no .Renviron file can be
found, one will be created. \strong{Warning:} only call \code{store_creds}
once to prevent multiple environment variables being stored.

If you do not want to alter your global .Renviron file, then use
\code{store_creds_temp} which stores environment variables for only the
current R session. Alternatively an API token and user agent can be passed as
arguments directly into the relevant newsrivr functions. This method is not
recommended as credentials can accidentally get leaked in scripts and .Rhistory
files.
}

\examples{
\dontrun{
store_creds() # you will be prompted to enter your credentials

store_creds_temp() # you will be prompted to enter your credentials
}

}
